requirements.txt
alembic==1.0.5
autopep8==1.4.3
googleapis-common-protos==1.5.5
grpcio==1.17.1
grpcio-tools==1.17.1
Mako==1.0.7
MarkupSafe==1.1.0
mysql==0.0.2
mysqlclient==1.3.14
protobuf==3.6.1
pycodestyle==2.4.0
PyMySQL==0.9.3
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
virtualenv==16.2.0
